New Navbar, Updated Sir Trevor, Started Implmentation of Image Blcok
The previous navbar was causing problems. The new navbar is Twitter Bootstrap's native nabar.
This means it's responsive, looks pretty and works (no custom JS is necessary) and is
customized in < 4 lines of code. If you have any problems with it, check out Bootstrap's docs -
they're very detailed. SirTrevor is the block editor we're using - there's a hidden branch
that contains an updated version. The old version had a bunch of random JS errors, with poor
documentation, and didn't work perfectly. This one is slightly better. With the new version, it
greatly simplifies image blocks.