Image Upload Is Now Working!
Stupid f***** issue where the image filename changed twice: once
when you uploaded the image (expected, as we remove the filename
user's provided to prevent duplicates) but then it was also
changing the filename after the user saved the Image. This was
really confusing because nowhere did I randomize it twice. Turns
out, that you need to store/cache the randomly generated name
to look it up for that second time. Before, we were just generating
a new one, but now we look to see if it's already been set and if
it has then we just use that one.....meaning it renames it only once!