Added Title Helper Method, Fixed JS and Started Preparing For Production
There is a title method now that can be called at the top of any view
by:
	<%= title "ENTER_TITLE_HERE" %>

Also added a new stripped down seed file that simply creates the
categories and subcategories (no fake data is provided). This will
be useful when uploading the project to Hopper - we will be able to
run this to create all the sample data.

Next, I moved someone's copy and pasta'd JS from the header file to
a proper file in the Asset Pipeline. I also removed the error they
had in it.

As always, there's also a bit of cleaning up code.